% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{bin_integer_fct}
\alias{bin_integer_fct}
\title{Bin integer levels in a factor}
\usage{
bin_integer_fct(ff, cuts)
}
\arguments{
\item{ff}{Factor vector to be binned.}

\item{cuts}{A numeric vector giving bin cuts. Don't add \code{-Inf} or \code{Inf} end cuts as these are handled automatically.}
}
\description{
Input is a vector with levels that, all except one, are naturally integers e.g. "60". The other level is a
string e.g. "missing".  An ordered factor is returned with the "integer" levels grouped into bins
based on the argument \code{cuts} with appropriate names.
}
\details{
FACTOR ORDERING HAS NOT BEEN CHECKED.
}
