% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{bin_numeric}
\alias{bin_numeric}
\title{Depreciated. Use \code{bin_integer()} instead.}
\usage{
bin_numeric(x, cuts, na_level = NULL, na_at_end = FALSE)
}
\description{
Depreciated. Use \code{bin_integer()} instead.
}
