% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_sets}
\alias{compare_sets}
\title{Compare two sets for overlap and differences.}
\usage{
compare_sets(x, y, summary = TRUE)
}
\arguments{
\item{x, y}{Vectors to compare.}

\item{summary}{Logical, controls what information is returned. See details.}
}
\description{
Compares two vectors and returns information on their overlap and differences, ignoring
ordering and duplicated elements.
}
\details{
A tibble is returned with contents dependent on the argument \code{summary}. If \code{TRUE},
counts of elements in both sets and only one set is given. If \code{FALSE}, the returned tibble
will have a row for each element in the union of \code{x} and \code{y}, together with which set(s)
it belongs to.
}
