% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_vars}
\alias{compare_vars}
\title{Pairwise element comparison of two variables in a dataframe}
\usage{
compare_vars(df, ..., simple = FALSE, tol = 0.000001, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{...}{The two columns to be compared either by name (quoted or unquoted) or integer positions.}

\item{simple}{Logical. If TRUE then only checks for pairwise differences, not ordering.}

\item{tol}{When comparing numeric vectors, equality and inequality comparisons use this tolerance.}

\item{na.rm}{If \code{TRUE}, will ignore rows where either element is \code{NA}. The main reason to use
this is if you want to see the proportion of non-missing entries that match/differ.}
}
\description{
Returns a table summarising the ordering of corresponding elements of two columns in a data frame. Columns
can be of different type or class but will then depend on coercion.
}
