% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{convert_date}
\alias{convert_date}
\title{Parse a string in a particular format and convert to a date or datetime.}
\usage{
convert_date(x)
}
\arguments{
\item{df}{A data frame.}

\item{var1, var2}{String names of the two columns to be compared.}
}
\description{
Reads text vectors which are either in dmy or dmy:time format. If all of the times are
zeroes then the output is a date. Fully `NA` vectors are returned as dates.
}
