% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-all.R
\name{count_at}
\alias{count_at}
\title{Perform  \code{dplyr::count} for a range of variables in a data frame (DEPRECATED - use \code{count_over()})}
\usage{
count_at(df, cols = NULL, sort = TRUE, n = 10)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{Vector of integer indices. If missing, all columns are included.}

\item{sort}{Logical passed to \code{count()} to say whether results are sorted by descending number of observation.
Unlike in \code{count()}, this defaults to \code{TRUE}.}

\item{n}{Integer passed to \code{print()} which gives the maximum number of rows printed in each count summary.}
}
\description{
Prints output from \code{dplyr::count()} for each variable index given by argument \code{cols} (an integer vector).
}
