% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_levels}
\alias{count_levels}
\title{count_nas
Counts NAs in a tibble by column.}
\usage{
count_levels(tbl)
}
\arguments{
\item{tbl}{a tibble.}
}
\description{
Returns a vector of the number of levels of each variable in a tibble.
}
