% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_matches2}
\alias{count_matches2}
\title{Count exact string matches in a data frame by column.}
\usage{
count_matches2(df, strings)
}
\arguments{
\item{df}{A data frame.}

\item{strings}{A character vector.}
}
\description{
Similar to \code{count_matches()} but counts matches for multiple strings rather than just one. The
output is a tibble with a row for each column in \code{df}. Unlike \code{count_matches}, only non-string
matching is not enabled.
}
