% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_n}
\alias{count_n}
\title{Count group sizes by grouped variables}
\usage{
count_n(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Variables to group by.}
}
\description{
Shorthand for \code{count(count(df, ...), n)}.
}
\examples{
count_n(mtcars, disp)

}
