% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_nas}
\alias{count_nas}
\title{count_nas: counts NAs in a tibble by column.}
\usage{
count_nas(tbl, prop = F)
}
\arguments{
\item{tbl}{a tibble.}

\item{prop=F}{Set to true to return the number of NAs as a proportion of rows, and count otherwise.}
}
\description{
Returns a vector of the number or proportion of NAs of each variable in a tibble.
}
