% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_nas2}
\alias{count_nas2}
\title{count_nas2: counts NAs in a data frame by column.}
\usage{
count_nas2(df, all = FALSE, sort = TRUE)
}
\arguments{
\item{df}{a data frame.}

\item{all}{By default variables with no NAs are omitted from the output. Set to TRUE to show all.}

\item{sort}{By default the output table is sorted by descending number of NAs. Set to FALSE to keep
variable ordering as in the data.}
}
\description{
Unlike `count_nas()` this returns the results in a tibble with columns giving:

* The number of NAs for each variable in the data.
* The number of NAs as a proportion of rows.
* The class of each variable.
}
