% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_over}
\alias{count_over}
\title{Perform  \code{count2()} for a range of variables in a data frame}
\usage{
count_over(df, ..., sort = TRUE, n = 10L)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Columns to count over. Accepts tidyselect inputs. If omitted then count is applied to every column.}

\item{sort}{Logical, passed to \code{count()}. If \code{TRUE} each result is sorted by group size.}

\item{n}{Integer passed to \code{print()} which gives the maximum number of rows printed in each count summary.}
}
\description{
Prints output from \code{count2()} for multiple columns in \code{df}.
}
