% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_unique}
\alias{count_unique}
\title{Counts NAs in a tibble by column.}
\usage{
count_unique(tbl)
}
\arguments{
\item{tbl}{a tibble.}
}
\description{
Returns a vector of the number of unique values of each variable in a tibble.
}
