% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{dir_contents}
\alias{dir_contents}
\alias{dir_size}
\title{Summarise the contents of a directory}
\usage{
dir_contents(dir = ".", ...)

dir_size(path = ".")
}
\arguments{
\item{dir}{A character vector of one or more directory paths.}

\item{...}{Additional arguments passed to \code{dir_info()}.}
}
\value{
The data frame returned by \code{dir_contents()} has columns as follows.
\item{directory}{The path of the directory.}
\item{files_size}{The total size of all files the directory, as a \code{fs::fs_bytes()} numeric vector.}
\item{num_files}{The number of files the immediate directory.}
\item{num_dirs}{The number of immediate sub-directories of the directory.}
\item{total_size}{The total size of all files the directory, including sub-directories.}
}
\description{
The main purpose of these functions is to give information about the size of directories.
\code{dir_contents()} returns a data frame with rows for each directory or sub-directories.
\code{dir_size()} returns a \code{fs::fs_bytes()} numeric vector of the total size the directories, including
the contents of sub-directories.
}
