% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{dir_files}
\alias{dir_files}
\title{Simple summary of files in a directory}
\usage{
dir_files(dir = ".", ...)
}
\arguments{
\item{dir}{A character vector of one or more directory paths.}

\item{...}{Additional arguments passed to \code{dir_info()}.}
}
\value{
A data frame with columns as follows.
\item{file_name}{The name of the file, as a character vector.}
\item{ext}{The extension of the file, as a character vector.}
\item{size}{The file size as an \code{\link[=fs_bytes]{fs_bytes()}} vector.}
\item{mod_date}{The date of last data modification.}
\item{dir}{The directory part of the file path, as a character vector.}
}
\description{
Similar to \code{fs::dir_info()} but with reduced output. Only files, not sub-directories are included.
}
