% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edwards-package.R
\docType{package}
\name{edwards-package}
\alias{edwards}
\alias{edwards-package}
\title{edwards: James Edwards personal package}
\description{
A collection of miscellaneous data science functions for my own use.
}
\keyword{internal}
