% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{err_by_class}
\alias{err_by_class}
\title{Error rate by class prediction.}
\usage{
err_by_class(pred_vec, target, class = 1)
}
\arguments{
\item{pred_vec}{Integer vector of class predictions.}

\item{target}{Integer vector of true classes.}

\item{class}{Integer indicating which class predictions to measure.}
}
\description{
Returns the rate of mismatches between entries of `pred_vec` that equal `class` and the
corresponding elements of `target`.
}
