% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_vector}
\alias{extend_vector}
\title{Extend vector to given length}
\usage{
extend_vector(x, n)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Length one numeric.}
}
\value{
A vector of length \code{n}.
}
\description{
Appends \code{NA} values to \code{x} so that its length is \code{n}.
}
\examples{
extend_vector(2:5, 7)
extend_vector(2:5, 2)
}
