% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Convert a factor with numeric or logical levels to a numeric vector}
\usage{
factor_to_numeric(x)
}
\arguments{
\item{x}{A factor.}
}
\description{
The obvious \code{as.numeric()} is incorrect. No checks are made on the level contents except to check if levels are
\code{c("TRUE", "FALSE")}, in which case they will be converted to a integer vector (0 for FALSE, 1 for TRUE).
}
\details{
Numeric part is from https://stackoverflow.com/questions/3418128/how-to-convert-a-factor-to-integer-numeric-without-loss-of-information
}
