% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{find_similar_single}
\alias{find_similar_single}
\title{Count pairwise matches between columns in a single data frame}
\usage{
find_similar_single(df)
}
\description{
Is called from \code{find_similar()} when the second data frame argument is omitted.
See \code{find_similar()} documentation for details.
}
\keyword{internal}
