% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet_to_table}
\alias{glmnet_to_table}
\title{Summarise coefficients from glmnet in a table}
\usage{
glmnet_to_table(fit, var_names = NULL, s = "lambda.1se", min_coef = 1e-10)
}
\arguments{
\item{fit}{A fitted glmnet model.}

\item{var_names}{(optional) A character vector of column names for data used in \code{fit}.}

\item{s}{The regularisation parameter determines which model is used from \code{fit} (as used in glmnet).}

\item{min_coef}{Coefficients with smaller absolute value than this are excluded from the table.}
}
\description{
Returns a tibble of coefficients for glmnet model \code{fit} with parameter \code{s}. Only
coefficients with absolute value greater than \code{min_coef} are included. If \code{var_names} is
supplied then a column of "parents" for each coefficient will be added. The parent is the data column
from which the variable or factor level is taken (first match in the case of interactions).
}
