% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ilogit}
\alias{ilogit}
\title{Vectorised inverse logit function: $exp(x) / (1 + exp(x))$.}
\usage{
ilogit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Modified inverse logit function from Faraway package. The original allowed for some elements to be NA which
  this does not. I added a check on `x` to avoid `NaN`. This occurs when x >~750 but
  since exp(x) / (1 + exp(x)) = 1 for x >=20 I only check for x > 20 and output 1 for these cases.
}
