% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{is_one2one}
\alias{is_one2one}
\title{Check data frame columns for one-to-one relationship.}
\usage{
is_one2one(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Two or more columns to be compared either by name (quoted or unquoted) or integer positions.}
}
\description{
Returns \code{TRUE} if all chosen columns are one-to-one and \code{FALSE} otherwise.
A message gives the first column found with a many-to relationship with another column.
}
