% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kbl_ctable}
\alias{kbl_ctable}
\title{\code{edwards::my_kable()} with a two way contingency table input.}
\usage{
kbl_ctable(ct, digits = 3)
}
\arguments{
\item{ct}{A two way table object.}

\item{digits}{Sets the number of digits (passed to \code{my_kable()}).}
}
\description{
\code{edwards::my_kable()} with a two way contingency table input.
}
