% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{latest_file}
\alias{latest_file}
\title{Get file name of most recent file in a directory}
\usage{
latest_file(path, root_name)
}
\arguments{
\item{path}{A string path name for the directory to search in.}

\item{root_name}{A string giving the part of the file name before the date.}
}
\description{
Looks for files in a directory \code{path} with names beginning \code{"root_name_dd"}, where
dd are digits, then returns the full path of the file with the maximum numeric part of its name.
This maximum is decided alphabetically so the numeric part of the file names should be structured
similarly across the compared filed e.g. a two digit version number or date in yyyy-mm-dd form.
}
