% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_n}
\alias{min_n}
\title{nth smallest element in a vector}
\usage{
min_n(x, n = 2L)
}
\arguments{
\item{x}{A vector.}

\item{n}{Numeric vector giving the ranks of elements to be returned.}
}
\description{
nth smallest element in a vector
}
