% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{mode_stat}
\alias{mode_stat}
\title{Calculate the statistical mode of an atomic vector}
\usage{
mode_stat(x, multiple = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{An atomic vector.}

\item{multiple}{Boolean. If TRUE returns a vector of modes, or just the first otherwise.}

\item{na.rm}{Boolean. Indicates whether to ignore NAs.}
}
\description{
Returns the statistical mode of an atomic vector (can be numeric, character or factor).
}
\details{
From https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode.
}
