% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{my_kable}
\alias{my_kable}
\title{Version of \code{kable()} with standard styling}
\usage{
my_kable(df, digits = 3)
}
\arguments{
\item{df}{A data frame.}

\item{digits}{Sets the number of digits (via the \code{kable()} \code{digits} argument).}
}
\description{
To save typing. Returns a table for R Markdown with \code{kableExtra::kable_styling}
options \code{bootstrap_options = c("striped", "condensed")} and \code{full_width = F}.
The number of digits is set using the argument \code{digits} (defaults to 3).
}
