% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_if_string.R
\name{na_if_any}
\alias{na_if_any}
\title{Convert values to NA}
\usage{
na_if_any(x, y)
}
\arguments{
\item{x}{Vector to modify.}

\item{y}{Values to replace with NA.}
}
\value{
A modified version of \code{x} that replaces any values that are equal to any element of y with NA.
}
\description{
Convert values in \code{x} to \code{NA} if they match any elements of \code{y}.
}
\examples{
na_if_any(1:5, c(2, 4))
}
\seealso{
This is similar to \code{dplyr::is_na()} but checks each element of \code{x} for a match with
any elements of \code{y}.
}
