% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{need}
\alias{need}
\title{Check if a package is installed and stop if not}
\usage{
need(package)
}
\arguments{
\item{package}{a package name as a character string.}
}
\description{
If the package is not installed then the function will error and give an appropriate message.
}
\details{
Designed to be used in scripts or functions where a function from the named package is used with \code{::}
without attaching the package. It should be used alongside \code{library()} calls at the top
of scripts where it also serves the purpose of indicating to the user that a package is needed.

The function uses \code{requireNamespace()} which will load the package if available but not attach it.
This is what happens when \code{::} is used.
}
