% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negation of \code{\%in\%}}
\usage{
a \%notin\% b
}
\arguments{
\item{a, b}{Vectors passed to \code{\%in\%}.}
}
\description{
Binary operator, which returns a logical vector with \code{TRUE} element where there is
no match in the left operand and \code{FALSE} where there is a match.
}
