% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{object_size_all}
\alias{object_size_all}
\title{object_size_all: returns table of sizes of all object in Mb, in order}
\usage{
object_size_all()
}
\value{
A tibble with two columns: object name and size in Mb.
}
\description{
object_size_all: returns table of sizes of all object in Mb, in order
}
