% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{oob_pred}
\alias{oob_pred}
\title{Helper for rang_oob_err.}
\usage{
oob_pred(pred_mat, n_trees)
}
\arguments{
\item{pred_mat}{Matrix of out-of-bag individual tree class predicitions with a zero where the prediciton
was in bag. Rows are observation, columns are trees.}

\item{n_trees}{Positive integer. Gives the number of trees to be used in the prediction (columns `1:n_trees`
of `pred_mat` will be used).}
}
\description{
Returns vector of class predictions chosen by majority vote.
}
\details{
Note predictions may include NaNs if there are no out-of-bag predictions for a data point.
}
