% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-plot.R
\name{plot_nas}
\alias{plot_nas}
\title{Plot missing values per variable (depreciated)}
\usage{
plot_nas(df, show_all = FALSE, sort = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{show_all}{Logical. Default FALSE doesn't show variables with no NAs.}

\item{sort}{Logical. Sort variables by number of NAs? Defaults to \code{TRUE}.}
}
\description{
Based on \code{DataExplorer::plot_missing} but simpler.
}
