% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop_ci}
\alias{prop_ci}
\title{Confidence intervals for binary target variable by values of a discrete predictor.}
\usage{
prop_ci(dt, target_name, var_name, min_n = 1, show_all = T,
  conf.level = 0.95, prop_lim = NULL)
}
\arguments{
\item{dt}{A data frame.}

\item{target_name}{String. Name of target variable. Must be binary (either TRUE/FALSE or 0/1).}

\item{var_name}{String. Name of predictor variable.}

\item{min_n}{Integer >= 1. Predictor levels with less than `min_n` observations are not displayed.}

\item{show_all}{Boolean. Defaults to TRUE. If FALSE will not show levels whose confidence interval
overlaps the mean response of all observations.}

\item{conf.level}{Numeric in (0,1). Confidence level used for confidence intervals.}

\item{prop_lim}{Optional x axis limits passed to `ggplot()` e.g. `c(0,1)`.}
}
\description{
For a data frame input, one variable is a binary target (`target_name`) and another is selected to
be a predictor variable (`var_name`). Mean response and a confidence interval is calculated for the
target variable for each level or value of the predictor. The results are plotted and returned as a
table. The function most appropriate for factor predictors but will work with other variable types also.
}
\details{
Currently requires various functions from various tidyverse packages.
* Different colours are used depending on whether the interval is above/below/overlapping the population
  mean. Currently sometimes inconsistent.
* Predictor levels are sorted by n but good to add option to keep usual ordering. Detect ordered factors?
* Doesn't handle NAs currently.
* Target must currently be a non-factor. Need to handle this.
}
