% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang_oob_err}
\alias{rang_oob_err}
\title{Out-of-bag error rates by number of trees for a ranger random forest}
\usage{
rang_oob_err(rf, data)
}
\arguments{
\item{rf}{A ranger random forest object.}

\item{data}{A data frame used to fit `rf`.}
}
\description{
Returns a table of out-of-bag error rates for a ranger randomw forest using number of trees
from 10 to all trees in steps of 10.
}
