% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang_roc_cut}
\alias{rang_roc_cut}
\title{ROC curve optimal cut for a ranger object}
\usage{
rang_roc_cut(rf, target, plot = TRUE)
}
\arguments{
\item{rf}{A ranger fitted model.}

\item{target}{A binary class target vector matching `rf`.}

\item{plot}{Produce a plot. Defaults to `TRUE`.}
}
\description{
Calls `roc_cut()` for a fitted ranger model. See `roc_cut()` for details. Assumes `probability = TRUE` was
used when the ranger model was fitted.
}
