% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc_cut}
\alias{roc_cut}
\title{Optimal cut for ROC curve}
\usage{
roc_cut(perf, pred, plot = T)
}
\arguments{
\item{perf}{A ROCR performance object.}

\item{pred}{A ROCR prediction object.}

\item{plot}{Logical indicating whether to plot a ROC curve together with distance from the optimal corner.}
}
\description{
Calculates the "optimal" cutoff for a classifier from a ROC curve. This is the cutoff that minimises the
distance from the point (FPR = 0, TPR = 1). Also returns sensitivity and specitivity for this cutoff and
optionally plots the curve together with the distance measure.
}
\details{
Adapted from https://www.r-bloggers.com/a-small-introduction-to-the-rocr-package/.
}
