% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{save_check}
\alias{save_check}
\title{Conditionally save file to RDS.}
\usage{
save_check(object, file, overwrite = FALSE, ...)
}
\arguments{
\item{object}{Object to save.}

\item{file}{The name of the file where object is to be saved.}

\item{overwrite}{Will only overwrite existing file if set to \code{TRUE}.}

\item{...}{Other arguments passed to \code{saveRDS()}}
}
\value{
(invisible) The \code{file} argument.
}
\description{
Wrapper around \code{saveRDS()} but checks if file already exists first. If it does
then it will only save if \code{overwrite = TRUE}. A message is given whether the object is
saved or not.
}
