% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{split_kable}
\alias{split_kable}
\title{Split a table in into parts of balanced size}
\usage{
split_kable(df, index, n_tb = 2, digits = 3)
}
\arguments{
\item{df}{A data frame.}

\item{index}{Which part of the split table to return.}

\item{n_tb=2}{The number of parts to split the table into.}

\item{digits=3}{Sets the number of digits (via the `kable()` `digits` argument).}
}
\description{
Splits a table into a `n_tb` parts of similar sizing using standard styling in `my_kable()`. This is
  used when I want to split a long thin table across multiple columns to reduce a document's size. Returns
  a single table part. Which part given by argument `index`.
}
\details{
Printing all table parts in one function call didn't work since I couldn't get
  `kable()` to print from within a function.
}
