% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matches.R
\name{string_missing}
\alias{string_missing}
\title{Strings commonly indicating missing values}
\usage{
string_missing()
}
\value{
A character vector \code{c("", " ", ".", "-", "NA", "na", "n/a", "N/A", "(missing)")}.
}
\description{
Used as a default argument in \code{count_matches()} and \code{var_summary()}.
}
