% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{vcount}
\alias{vcount}
\alias{countv}
\title{Frequency count for values of a vector with tibble output.}
\usage{
vcount(x, sort = TRUE, name = NULL, value_name = "value")

countv(x, sort = TRUE, name = NULL, value_name = "value")
}
\arguments{
\item{x}{An atomic vector.}

\item{sort}{Logical. Sort output by descending frequency. If \code{FALSE} sort by value.}

\item{name}{Name of count column in output. If omitted \code{n} will be used (as in \code{dplyr::count}).}

\item{value_name}{Name of value column.}
}
\description{
This behaves similarly to \code{count2()} but with a vector input. \code{countv()} is an alternative name.
}
