% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{bin_numeric}
\alias{bin_numeric}
\title{Converts numeric into binned factor with appropriately named levels.}
\usage{
bin_numeric(x, cuts, na_level = NULL, na_at_end = FALSE)
}
\arguments{
\item{cuts}{A numeric vector giving bin cuts. Don't add `-Inf` or `Inf` end cuts as these are handled automatically.}

\item{na_level}{A name for the explicit NA level (if any). If left as the default NULL then no explicit NA
level will be formed.}

\item{na_at_end}{Boolean indicating whether to put any explicit NA level at the end of the levels. The
default `na_at_end = FALSE` will put the level at the start.}
}
\description{
Takes a numeric input vector, groups into bins based on the argument `cuts`, and
outputs an ordered factor with appropriate names. Any NAs in the input are optionally grouped
into an explicit NA level with name `na_level`
}
