% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count2}
\alias{count2}
\title{\code{dplyr::count()} with proportion column and default \code{sort = T}.}
\usage{
count2(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Arguments passed to count including variables to group by.}
}
\description{
Adds a column "prop" which gives the proportion of total rows in that group.
}
\examples{
count2(mtcars, gear)

}
