% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count2}
\alias{count2}
\alias{countp}
\title{\code{dplyr::count()} with proportion column and default \code{sort = TRUE}.}
\usage{
count2(df, ..., sort = TRUE, wt = NULL, name = NULL, pname = "prop")

countp(df, ..., sort = TRUE, wt = NULL, name = NULL, pname = "prop")
}
\arguments{
\item{df}{A data frame.}

\item{...}{Variables to group by. Passed to \code{count()}.}

\item{sort}{Passed to \code{count}, but defaults to \code{TRUE}.}

\item{wt, name}{Optional, passed to \code{count}.}

\item{pname}{Name for the proportion column.}
}
\description{
Adds a column \code{pname} which gives the proportion of total rows in that group.
\code{countp()} is an alternative name.
}
\examples{
count2(mtcars, cyl)
count2(mtcars, cyl, sort = FALSE)

}
