% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_at}
\alias{count_at}
\title{Performs dplyr::count for a range of variables in a data frame.}
\usage{
count_at(df, cols = NULL, sort = TRUE, n = 10)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{Vector of integer indices. If not given counts for all columns are given.}

\item{sort}{Boolean passed to `count()` to say whether results are sorted by descending number of observation..
Defaults to TRUE (unlike `count()`).}

\item{n}{Integer passed to `print()` which gives the maximum number of rows printed in each count summary.}
}
\description{
Prints output from dplyr::count for each variable index given by argument `cols` (an integer vector).
}
