% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_summary.R
\name{count_levels}
\alias{count_levels}
\title{Count the number of levels in a data frame by column}
\usage{
count_levels(df, all = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{all}{By default variables with no levels are omitted from the output. Set \code{all=TRUE} to show all.}
}
\description{
Returns a vector of the number of levels of each variable in a data frame.
}
