% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_matches}
\alias{count_matches}
\title{count_matches: Counts the total number of exact matches to a value in a data frame by column.}
\usage{
count_matches(df, value, all = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{value}{A length one vector.}

\item{all}{By default variables with no matches are omitted from the output. Set all=T to show all.}
}
\description{
Returns a named integer vector with elements that give the number of entries in the corresponding
column of \code{df} that match to the argument \code{value}. No coercion is used so type must also match.
}
\examples{
x <- data.frame(a = c("an", "banana", "candy"), b = c("on", "bon", "bonbon"), c = 1:3)
count_matches(x, "an", all = T)
count_matches(x, "an")
count_matches(x, 1L)
count_matches(x, 1) # type must match
count_matches(x, "1") # type must match

}
