% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_nas}
\alias{count_nas}
\title{Counts NAs in a data frame by column.}
\usage{
count_nas(df, all = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{all}{By default variables with no NAs are omitted from the output. Set to TRUE to show all.}
}
\description{
Returns a vector of the number of NAs of each variable in a data frame.
}
