% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_nas2}
\alias{count_nas2}
\title{Count NAs in a data frame by column (depreciated)}
\usage{
count_nas2(df, all = FALSE, sort = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{all}{By default variables with no NAs are omitted from the output. Set to \code{TRUE} to show all.}

\item{sort}{By default the output table is sorted by descending number of NAs. Set to \code{FALSE} to keep
variable ordering as in the data.}
}
\description{
Unlike \code{count_nas()} this returns the results in a tibble with columns giving:
\itemize{
\item{The number of NAs for each variable in the data.}
\item{The number of NAs as a proportion of rows.}
\item{The class of each variable.}
}
}
