% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_string}
\alias{count_string}
\title{Counts the total number of string pattern matches in a data frame by column.}
\usage{
count_string(df, pattern, all = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{all}{By default variables with no levels are omitted from the output. Set all=T to show all.}
}
\description{
Returns a vector of the number of matches to a string pattern `pattern` of each variable in a data frame.
}
