% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_string}
\alias{count_string}
\title{Count the total number, by column, of entries in a data frame that match a string pattern}
\usage{
count_string(df, pattern, all = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{pattern}{A string pattern (regular expression).}

\item{all}{By default variables with no matches are omitted from the output. Set all=T to show all.}
}
\description{
Returns a named integer vector with elements that give the number of entries in the corresponding
column of \code{df} that contain a match to the string pattern \code{pattern}. No coercion is used
so only characters or factors are matched (see examples).
}
\details{
Note that repeated occurences of \code{pattern} in a single string are only counted once (see examples).
}
\examples{
x <- data.frame(a = c("an", "banana", "candy"), b = c("on", "bon", "bonbon"), d = 1:3)
count_string(x, "an", all = T)
count_string(x, "an")
count_string(x, "b")
count_string(x, "1") # not matched to integers

}
