% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_unique}
\alias{count_unique}
\title{Count the number of unique values in a data frame by column}
\usage{
count_unique(df)
}
\arguments{
\item{df}{A data frame.}
}
\description{
Returns a vector of the number of unique values of each variable in a data frame. Any NA entries are included as
  a unique value.
}
