% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{diff_days}
\alias{diff_days}
\title{Difference in days of two dates as a numeric}
\usage{
diff_days(date1, date2)
}
\arguments{
\item{date1}{A date or datetime.}

\item{date2}{A date or datetime.}
}
\description{
Finds \code{difftime(date1, date2)} using \code{units = "days"}, then converts to a numeric. If either
input is a datetime then only the date part is used (the datetime is first coerced to a date using
\code{lubridate::as_date()}).
}
