% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_contents.R
\name{dir_files}
\alias{dir_files}
\title{Simple summary of files in a directory}
\usage{
dir_files(dir = ".", ..., shorten = TRUE)
}
\arguments{
\item{dir}{Character. A single directory path.}

\item{...}{Additional arguments passed to \code{dir_info()}.}

\item{shorten}{If \code{TRUE} (default), the directory column gives the path relative to \code{dir} (a "."
is given when the directory is \code{dir}). If \code{FALSE}, the full path is given.}
}
\value{
A data frame with columns as follows.
\item{file_name}{The name of the file, as a character vector.}
\item{ext}{The extension of the file, as a character vector.}
\item{size}{The file size as an \link{fs_bytes} vector.}
\item{mod_time}{The date and time when the file was last modified.}
\item{directory}{The directory part of the file path, as a \code{fs::path} vector.}
\item{level}{The level of the directory relative to \code{dir}}.
}
\description{
Similar to \code{fs::dir_info()} but only files, not sub-directories are included, and with fewer
columns in the output. Unlike \code{fs::dir_info()}, \code{dir} must be length one.
}
