% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Convert a factor to a numeric vector using its levels}
\usage{
factor_to_numeric(x)
}
\arguments{
\item{x}{A factor.}
}
\description{
Converts to a numeric vector factors with levels that could be appropriately converted to
numeric values (either number or logical-type strings).
}
\details{
The output from the obvious \code{as.numeric()} is unlikely to be what is desired, since it acts
on the underlying integers, rather than the levels. This function instead uses
\code{as.integer(as.logical(x))} to attempt to coerce the character levels. If all levels are
strings that would be coercible by \code{as.logical()}, that is in
\code{c("T", "TRUE", "True", "true", "F", "FALSE", "False", "false", NA)}, then they
are converted appropriately to a numeric vector instead. The two methods are never mixed, so
\code{factor_to_numeric(factor(c("T", "1")))} will attempt numeric conversion and produce \code{NA}
for the \code{"T"}.
}
\examples{
x <- factor(3:5)
as.numeric(x)
factor_to_numeric(x)
factor_to_numeric(factor(c("T", "false", NA)))
}
