% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Convert a numeric factor to a numeric vector}
\usage{
factor_to_numeric(x)
}
\arguments{
\item{x}{A numeric factor.}
}
\description{
The obvious `as.numeric()` is incorrect.
}
\details{
From  https://stackoverflow.com/questions/3418128/how-to-convert-a-factor-to-integer-numeric-without-loss-of-information
}
