% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_to_table.R
\name{glmnet_to_table}
\alias{glmnet_to_table}
\title{Summarise coefficients from glmnet in a table}
\usage{
glmnet_to_table(fit, s = "lambda.1se", min_coef = 1e-10)
}
\arguments{
\item{fit}{A fitted glmnet model.}

\item{s}{The regularisation parameter determines which model is used from \code{fit} (as used in glmnet).}

\item{min_coef}{Coefficients with smaller absolute value than this are excluded from the table.}
}
\description{
Returns a tibble of coefficients for glmnet model \code{fit} with parameter \code{s}. Only
coefficients with absolute value greater than \code{min_coef} are included.
}
\details{
This is a simplified version of \code{jemodel::glmnet_to_table()} kept here to provide some backwards
compatibility. Longer term it may be removed from this package.
}
