% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-all.R
\name{glmnet_to_table}
\alias{glmnet_to_table}
\title{Summarise coefficients from glmnet in a table}
\usage{
glmnet_to_table(fit, ..., min_coef = 1e-10)
}
\arguments{
\item{fit}{A fitted glmnet model.}

\item{...}{Arguments passed to \code{coef()}. For \code{glmnet}, most commonly used will be \code{s} (see \code{predict.cv.glmnet()}).}

\item{min_coef}{Coefficients with smaller absolute value than this are excluded from the table.}
}
\description{
Returns a tibble of coefficients for glmnet model \code{fit} with parameter \code{s}. Only
coefficients with absolute value greater than \code{min_coef} are included.
}
\details{
This is a simplified version of \code{jemodel::glmnet_to_table()} kept here to provide some backwards
compatibility. Longer term it may be removed from this package.
}
