% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{latest_file}
\alias{latest_file}
\title{Get file name of most recent file in a directory}
\usage{
latest_file(
  path = ".",
  root_name = ".*",
  file_ext = NULL,
  verbose = TRUE,
  n = 1L
)
}
\arguments{
\item{path}{A string path name for the directory to search in.}

\item{root_name}{A string giving the part of the file name before the date.}

\item{file_ext}{An optional string giving a file extension which the returned file name must end in.}

\item{verbose}{Give detail about matches files and other files containing \code{root_name}.}

\item{n}{Integer. Option to return the nth latest file.}
}
\description{
Looks for files in a directory \code{path} with names beginning \code{"root_name_dd"}, where
dd are digits. Returns the full path of the file with the maximum numeric part of its name.
This maximum is decided alphabetically so the numeric part of the file names should be structured
similarly across the compared files e.g. a two digit version number or date in yyyy-mm-dd form.
}
\details{
The string \code{root_name} can be a regex pattern. Lower and upper case versions of `\code{file_ext}
will also be matched.
}
