% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{latest_file}
\alias{latest_file}
\title{Get paths of files using file name ordering}
\usage{
latest_file(
  path = ".",
  pattern = NULL,
  ext = NULL,
  n = 1L,
  decreasing = TRUE,
  ...
)
}
\arguments{
\item{path}{A string path name for the directory to search in.}

\item{pattern}{A string pattern to filter file names by. The pattern is only applied to the part
of the path following the final "/".}

\item{ext}{An optional string giving a file extension which the returned file name must end in.
The supplied extension is checked for an exact match (ignoring case) to the ".xxx" part of the
file name.}

\item{n}{Integer. Return the nth file path. This can have length > 1, in which case multiple file
paths are returned. An error will be thrown if there are insufficient matches.}

\item{decreasing}{Logical passed to \code{order()}. By default returns the (nth) largest value.}

\item{...}{Other arguments passed to \code{fs:dir_ls()}. The argument \code{type} is always set to \code{"file"}
and so must not be supplied. Arguments \code{regexp} or \code{glob} are applied to the full file path and
so affect output differently than \code{pattern}.}
}
\description{
Looks for files in a directory \code{path} with file name matching \code{pattern}. The path of the \code{n}th
file, ordered alphabetically by file name is returned. "File name" here refers to the part of
the path without any directory names i.e. the part following the final "/".

The intended use is with files names with matching structure with well-ordered numeric endings,
such as "yyyy-mm-dd" or fixed length numeric ids. Output may not be as intended if
comparing files with different structures.
}
