% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_n}
\alias{max_n}
\title{nth maximum element in a vector}
\usage{
max_n(x, n = 2L, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{n}{Numeric vector giving the ranks of elements to be returned.}

\item{na.rm}{A logical indicating whether missing values should be ignored.
If \code{FALSE} an \code{NA} value in any of the arguments will cause a value of \code{NA}
to be returned.}
}
\description{
nth maximum element in a vector
}
