% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_n}
\alias{max_n}
\title{nth largest element in a vector}
\usage{
max_n(x, n = 2L)
}
\arguments{
\item{x}{A vector.}

\item{n}{Numeric vector giving the ranks of elements to be returned.}
}
\description{
nth largest element in a vector
}
