% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{my_kable}
\alias{my_kable}
\title{my_kable: Version of kable with standard styling to save typing.}
\usage{
my_kable(df, digits = 3)
}
\arguments{
\item{df}{A data frame.}

\item{digits=3}{Sets the number of digits (via the `kable()` `digits` argument).}
}
\description{
Returns a table for R Markdown with kableExtra::kable_styling options `bootstrap_options = c("striped", "condensed")`
  and `full_width = F`. The number of digits is set using the argument `digits` (defaults to 3).
}
