% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_missing2}
\alias{plot_missing2}
\title{Plot missing values per variable}
\usage{
plot_missing2(
  data,
  group = list(Low = 0.05, OK = 0.4, High = 1),
  geom_label_args = list(),
  title = NULL,
  ggtheme = theme_gray(),
  theme_config = list(legend.position = c("bottom")),
  show_all = F
)
}
\arguments{
\item{data}{Input data.}

\item{group}{Missing profile band taking a list of group name and group upper bounds.}

\item{geom_label_args}{A list of other arguments to \code{geom_label()}.}

\item{title}{Plot title.}

\item{ggtheme}{\code{ggplot2} themes.}

\item{theme_config}{A list of configurations to be passed to \code{theme()}.}

\item{show_all}{Logical. If \code{FALSE} (default) then do not show variables with no NAs.}
}
\description{
Adaptation of \code{DataExplorer::plot_missing} but with option to not show variables with no NAs.
Also has different default lables for each group.
}
\details{
For details and main arguments see \code{DataExplorer::plot_missing}.
}
