% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nas}
\alias{plot_nas}
\title{Plots missing values per variable.}
\usage{
plot_nas(df, show_all = F, sort = T)
}
\arguments{
\item{df}{A data frame.}

\item{show_all}{Boolean. Default FALSE doesn't show variables with no NAs.}

\item{sort}{Boolean. Sort variables by number of NAs? Defaults to `TRUE`.}
}
\description{
Based on `DataExplorer::plot_missing` but simpler.
}
