% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang_mtry}
\alias{rang_mtry}
\title{Tune ranger models for mtry}
\usage{
rang_mtry(
  data,
  fmla,
  m_vec,
  train = NULL,
  seed = 1,
  importance = "impurity",
  num.trees = 500,
  respect.unordered.factors = T
)
}
\arguments{
\item{data}{A data frame.}

\item{fmla}{A formula for the model.}

\item{train}{Optional integer vector giving rows indices to be used in training set. Remaining rows are used for
validation. If default \code{train = NULL} is used then all data is used for training and there is no validation.}

\item{seed}{Integer. Random number seed used for fitting each model.}

\item{importance, num.trees, respect.unordered.factors}{Optional arguments passed to \code{ranger()}. Defaults are
\code{"impurity"}, \code{500} and \code{TRUE} respectively.}

\item{mvec}{Integer vector of values of tuning parameter \code{mtry}.}
}
\description{
Fits \code{ranger()} models for a given range of values of mtry. Output is a table and graph giving
errors for OOB training data and optional validation data.
}
