% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang_oob_err}
\alias{rang_oob_err}
\title{Out-of-bag error rates by number of trees for a ranger random forest}
\usage{
rang_oob_err(rf, data, start = 5L, by = 5L, plot = TRUE)
}
\arguments{
\item{rf}{A ranger random forest object.}

\item{data}{A data frame used to fit \code{rf}. Must contain target variable.}

\item{start, by}{Error rates are evaluated for number of trees from \code{start} to maximum number
of trees in steps of \code{by}.}

\item{plot}{Optional logical. Output a plot or not.}
}
\description{
Returns a table of out-of-bag error rates for a ranger randomw forest using number of trees
from 10 to all trees in steps of 10.
}
