% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang_roc_cut}
\alias{rang_roc_cut}
\title{ROC curve optimal cut for a ranger object}
\usage{
rang_roc_cut(rf, target, plot = TRUE)
}
\arguments{
\item{rf}{A ranger fitted model.}

\item{target}{A binary class target vector matching \code{rf}.}

\item{plot}{Produce a plot. Defaults to \code{TRUE}.}
}
\description{
Calls \code{roc_cut()} for a fitted ranger model. See \code{roc_cut()} for details. Requires
 \code{probability = TRUE} to be used when the ranger model was fitted.
}
