% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{string_missing}
\alias{string_missing}
\title{Strings commonly indicating missing values}
\usage{
string_missing()
}
\value{
A character vector \code{c("", " ", ".", "-", "NA", "na", "n/a", "N/A", "(missing)")}.
}
\description{
Strings commonly indicating missing values
}
