% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{var_summary}
\alias{var_summary}
\title{Simple summary of the variables in a data frame}
\usage{
var_summary(df)
}
\arguments{
\item{df}{A data frame.}
}
\description{
Returns a tibble with the names, class, number of unique values, and the number and proportion of
\code{NA}s for each variable in the data. If there are \code{NA} values then they are included as
a unique value.
}
