% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{var_summary}
\alias{var_summary}
\title{Simple summary of the variables in a data frame.}
\usage{
var_summary(df)
}
\arguments{
\item{df}{a data frame.}
}
\description{
Returns a tibble with the names, class, number of NAs, number of unique values, and the number of
  levels for each variable in the data. Any NA entries are included as a unique value.
}
